requirements.txt
nbdev
transformers
#toolformer  # from this repository, install manually
torch
dill>=0.3.0,<0.3.7
torchtyping
einops
openai
#wolfram-alpha-api
langchain
python-dotenv
pytest
jupyterlab
jupyterlab-git
trl @ git+https://github.com/huggingface/trl@main
pyparsing
tqdm
#pydot2
dvc[gs]
dvc[ssh]
dvc[s3]
ollama